test(integration): named the repository-events steps file more clearly
to avoid collision with the upcoming steps file for repository settings rather than events
