+++ test/integration/features/step_definitions/common-steps.mjs
@@ -3,8 +3,18 @@ import { Probot, ProbotOctokit } from 'probot'
 import any from '@travi/any'
 import { Before, When } from '@cucumber/cucumber'
 import settingsBot from '../../../../index.js'
-import { buildRepositoryCreatedEvent, buildRepositoryEditedEvent } from './repository-steps.mjs'
+import { buildRepositoryCreatedEvent, buildRepositoryEditedEvent } from './repository-events-steps.mjs'
 import { buildPushEvent } from './config-steps.mjs'
+
+export const repository = {
+  default_branch: 'master',
+  name: 'botland',
+  owner: {
+    name: 'bkeepers-inc',
+    login: 'bkeepers-inc',
+    email: null
+  }
+}
 
 async function loadInstance () {
   const probot = new Probot({
