+++ test/integration/features/step_definitions/milestones-steps.mjs
@@ -7,7 +7,8 @@ import assert from 'node:assert'
 import any from '@travi/any'
 
 import settings from '../../../../lib/settings.js'
-import { repository } from './repository-steps.mjs'
+
+import { repository } from './common-steps.mjs'
 
 Given('no milestones exist', async function () {
   this.server.use(
